Classes, sections & courses views refractored

+++ app/Repositories/ProfileRepository.php
@@ -27,7 +27,10 @@ class ProfileRepository implements ProfileContract
         if(is_null($picture)) {
             $status = $profile->update($request->validated());
         } else {
-            Storage::disk('public')->delete($profile->picture);
+            if (!is_null($profile->picture)) {
+                Storage::disk('public')->delete($profile->picture);
+            }
+
             $path = $picture->store('users', 'public');
             $status = $profile->update([
                 'last_name' => $validated['last_name'],
